Dissolved `select_constraints.h` into per-feature AST headers
The header had grown to 803 lines holding six unrelated grammar features. Each now lives
in `dev/ast/`, grouped by feature the way `ast/window.h` is, rather than one node per file:

- `rowset_deduplicators.h` - DISTINCT, ALL
- `result_columns.h` - `as_optional()`, `columns()`, `struct_()`, `asterisk()`, `object()`
- `case_expression.h` - CASE, in both its simple and its searched form
- `select.h` - `select_t` and `validate_select_clauses()`, which states a rule about a
  select's clause pack and so belongs next to the node, not in the vocabulary layer
- `compound_operator.h` - UNION, UNION ALL, EXCEPT, INTERSECT
- `cte.h` - the CTE binding, the WITH clause and the materialization hints. A CTE is only
  ever reached through a WITH, so what distinguishes them are the axes (`is_cte_binding`
  vs `is_with_clause`), which a file boundary would not have added to.

Consumers match by node trait now instead of by concrete node type, which lets them drop
the node headers altogether: `ast_iterator.h` and `statement_serializer.h` include none of
the six, `column_expression.h` only the one whose node it destructures.

New vocabulary: `is_cte_binding`, `is_materialization_hint`, `is_case_expression` and
`is_distinct` on the grammar axis, `is_as_optional` on the structural axis, and a
`return_type_t` projection. `is_distinct` is classified on its own next to the
`is_rowset_deduplicator` group because DISTINCT carries a second role - it deduplicates
the argument rows of an aggregate function, which ALL does not.

`materialized_t` and `not_materialized_t` gained an implicit `std::string_view`
conversion, following `distinct_string` and `union_base`, so a single trait-matched
serializer replaces the two full specializations.

Left naming their concrete nodes: `node_tuple`'s destructuring specializations, and the
two `asterisk_t` cases in `column_expression.h` - each discriminates on the node's
operand, which is what matching the node itself is for.
