+++ dev/cte_storage.h
@@ -15,7 +15,9 @@
 #include "schema/column.h"
 #include "schema/table_base.h"
 #include "ast/quoted_expression.h"
-#include "select_constraints.h"
+#include "ast/result_columns.h"
+#include "ast/select.h"
+#include "ast/cte.h"
 #include "alias.h"
 #include "cte_types.h"
 #include "cte_column_names_collector.h"
