+++ dev/node_definitions.h
@@ -28,8 +28,11 @@
 
 #include "ast/quoted_expression.h"
 #include "ast/between.h"
+#include "ast/case_expression.h"
 #include "ast/cast.h"
+#include "ast/compound_operator.h"
 #include "ast/cross_join.h"
+#include "ast/cte.h"
 #include "ast/excluded.h"
 #include "ast/exists.h"
 #include "ast/group_by.h"
@@ -41,9 +44,11 @@
 #include "ast/match.h"
 #include "ast/offset.h"
 #include "ast/rank.h"
+#include "ast/result_columns.h"
+#include "ast/rowset_deduplicators.h"
+#include "ast/select.h"
 #include "ast/set.h"
 #include "ast/special_keywords.h"
 #include "ast/upsert_clause.h"
 #include "ast/where.h"
 #include "ast/window.h"
-#include "select_constraints.h"
