+++ dev/storage.h
@@ -36,7 +36,8 @@
 #include "alias.h"
 #include "field_printer.h"
 #include "operators.h"
-#include "select_constraints.h"
+#include "ast/result_columns.h"
+#include "ast/cte.h"
 #include "core_functions.h"
 #include "conditions.h"
 #include "statement_binder.h"
